pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "awesome-claude-md"
version = "1.0.0"
description = "A curated collection of high-quality claude.md files from public GitHub repositories"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "josix", email = "josix@users.noreply.github.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "requests==2.31.0",
    "PyGithub==1.59.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/josix/awesome-claude-md"
Repository = "https://github.com/josix/awesome-claude-md"
Issues = "https://github.com/josix/awesome-claude-md/issues"

[project.scripts]
discover-claude-files = "scripts.discover_claude_files:main"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a15",
    "complexipy>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"